requirements.txt
azure-functions
azure-identity
azure-storage-blob
azure-storage-file-datalake
requests
Pillow

